Split the feature context file out into a subcontext file
